+++ src/main/kotlin/com/nylas/resources/Events.kt
@@ -27,6 +27,23 @@ class Events(client: NylasClient) : Resource<Event>(client, Event::class.java) {
   }
 
   /**
+   * Returns a list of recurring events, recurring event exceptions, 
+   * and single events from the specified calendar within a given time frame. 
+   * This is useful when you want to import, store, and synchronize events 
+   * from the time frame to your application
+   * @param identifier Grant ID or email account to query
+   * @param queryParams The query parameters to include in the request (must include calendar_id)
+   * @param overrides Optional request overrides to apply
+   * @return The list of import events
+   */
+  @Throws(NylasApiError::class, NylasSdkTimeoutError::class)
+  @JvmOverloads
+  fun listImportEvents(identifier: String, queryParams: ListImportEventQueryParams, overrides: RequestOverrides? = null): ListResponse<Event> {
+    val path = String.format("v3/grants/%s/events/import", identifier)
+    return listResource(path, queryParams, overrides)
+  }
+
+  /**
    * Return an Event
    * @param identifier Grant ID or email account to query
    * @param eventId The id of the event to retrieve.